requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
